requirements.txt
Server
fastapi
uvicorn[standard]

Data Validation
pydantic

Environment Variables
python-dotenv

HTTP Requests
requests

LangChain Core and Integrations
langchain
langchain-google-genai
langchain-community

Vector Store and Embeddings
chromadb
sentence-transformers
huggingface-hub

Document Loader
pypdf